settings: Fix the Use Current Location button
During the Dolphin settings revamp
(https://invent.kde.org/system/dolphin/-/commit/489b56b68bb29e81337e115c490eea4403001b71?)
this QUrl had been forgot to add back so that the
FoldersTabsSettingsPage knows what is the current url.

This change checks for the main window of dolphin, then gets
the URL of it whenever useCurrentLocation is called.

However, when this is used as KCM, the button is not
created, since there it doesn't make sense.

BUG:491753