dolphintrash.cpp: Update trash dir on device changes
Make sure we update the trash directory when storage devices are added/changed.
This ensures that whenever a mounted device, for example
USB stick is mounted/unmounted, the contents are properly
updated in the trash folder as well.

Otherwise the trash folder and USB stick trash folder will get
out of sync, making the trash folder behavior unpredictable.

BUG:493247