Add test to avoid accidental divergence of desktop and phone ui
The dolphinui.rc and dolphinuiforphones.rc file describe parts of
the desktop and phone user interfaces for Dolphin. They are meant
to be different in specific ways, but identical in many others.

This test makes sure that the two ui files stay mostly identical by
requiring any differences between them need to be explicitly
stated/registered as an exception in the test. This way no
accidental divergence of user interfaces can happen.