KFileItemModel: fix `setData` in the renaming usage
The `setData` function is used by inline renaming to
update the `text` role (i.e. file name) in the model
before actually do the renaming.

However, the current implementation only updates `text`
and url in KFileItem in this case, leaving the `url` role
and the `m_items` cache untouched. This makes the cache
incoherent, so `index()` won't find the renamed item. It
also makes the `url` role incoherent with the `text` role.

This also fixes the problem mentioned in bug #481561
where when inline renaming fails, it still shows the new
name. (Because `m_model->index` fails to find the new name
and rename it back on job failure.)

BUG: 481561