dolphinview: after inline renaming, immediately update the selection
Currently, the inline renaming updates the model and
the list of selected files (m_selectedUrls) immediately
after the user presses enter, but the selection in the
view is not immediately updated. This results in the
selection to be briefly lost for about 1 second until
the view automatically refreshes.

This patch updates the selection immediately after
the user presses enter, so the renamed file is still
selected.

BUG: 481717