dolphinview: when rename dialog finishes, immediately update the model and the selection
On sucessful return of the rename dialog, we update
the model and the selection immediately to reflect
the new name. This is to avoid the short duration
after the rename during which the selection is lost.

Currently, after the rename dialog finishes, the selection
is briefly lost for about 1 second until the view
automatically refreshes.

This patch updates the model and selection immediately
after the dialog finishes, so the renamed file is still
selected.

BUG: 481717