Skip splitter in Tab order
When moving through Dolphin by repeatedly pressing the Tab key,
there is one seemingly invisible object that catches keyboard
focus. On a default toolbar it receives keyboard focus between the
"Details View Mode" button and the first button of the location
bar. That object is the splitter of the location bar itself, which
never requires manipulation by keyboard. This commit skips this
object by setting its focusProxy to the primaryUrlNavigator()
instead.