Test that each object has distinguishable accessible info
This commit extends the dolphinmainwindowtest for the accessible tree
to test that each object a user can reach through the Tab key or the
Shift+Tab key combination is distinguishable by its accessible info
from the previous object.

Objects are considered distinguishable if they have different
accessible names or their first ancestor with an accessible name is
different.