generalviewsettingspage: Use qobject_cast instead of static_cast
In generalViewSettingsPage, ActiveWindow is not necessarily a
mainWindow, which can cause dolphin to crash. For example when user
opens folder properties and then dolphin settings, this part thinks
the folder settings is mainWindow.

Instead of static_cast, use qobject_cast, since it will return nullptr
if the DolphinMainWindow is not found.

BUG:492741