DolphinView: beginAnchoredSelection after setSelected in selectNextItem
Currently when next item is set selected, it's not set as an anchor.
This means that shift-click will not select items between the selection
and the next item, and instead behaves like ctrl-click. 
This is rather inconsistent behavior.

We should set the next selected item as an anchor so that shift-click
will work: This change starts the `beginAnchoredSelection`
process after `setSelection`.

For example, if user trashes an item and the next item is selected,
shift-click would stop working. 

BUG:495371