DolphinTabPage: Update container view url on redirection
On url redirect, we should check which container url is being changed
and then update it accordingly. This makes sure the tab name is
updated.

We also should not disconnect redirection on view activation
since the redirection might be used by the other split. 
The disconnection is done in `setSplitViewEnabled` instead.

This allows us to update the tab name every time the url changes,
even inside a splitview where the split which name is changed is
not active.

BUG:496414