Make "open path" and "open path in new tab" scroll to the selected item
**Open Path**

When user clicks on "Open Path" after searching for an item,
user expects the view to show the item immediately.

We wait for the KItemListSmoothScroller to be done with its animation before the
scrollbar sizes are being changed.

**Open Path in New Tab**

When user selects "Open Path in New Tab", we open a new tab to the folder where the file is, then select and set the file current. We need to get the correct tab when opening one, so it has been added as a return value.

BUG:495613