Always focus the view after place activation
This one-liner makes sure keyboard focus always moves to the active
view when a place in the places panel is activated. Previously
focus would remain on the places panel if the activated location
was identical to the already displayed location.

This added consistency is helpful because keyboard-only users will
get used to be able to move from the places panel to the view by
pressing the Enter key, even though this does not always work
prior to this commit.

This issue was identified in an accessibility scan done by HAN
University of Applied Science.