ViewProperties: Return nullptr if viewPropertiesString is empty
If viewPropertiesString is empty, return a nullptr.
This will later used in the stack by the defaultProperties call.

In defaultProperties, if we can't find the global directory,
create new one with a tempfile. If tempfiles can't be created,
use default instead.

This will ensure that view settings are saved and loaded correctly
if user has separate view properties per folder.

This will also add an unit test, where we create a global directory,
modify it and make sure the changes are reflected in the unmodified
folder.

BUG:495878


(cherry picked from commit 8d155558e980197b4dc2180660ea9bc61a23f3d3)

97d9a70c ViewProperties: Return nullptr if viewPropertiesString is empty
60a3da35 Create lambda for creating temporary files

Co-authored-by: Akseli Lahtinen <akselmo@akselmo.dev>