Use logicalSize instead of cacheSize for preview overlays
Cachesize is wider than the size of the icon preview, so it
ends up pushing the overlay out of bounds of the icon size.
This causes the icon to shrink to make sure all the emblems fit.

For example, if icon is set to 64x64, but the overlay uses cacheSize
128x128, the preview image bounds are resized to 128x128. This
causes the preview image to shrink to make sure it can be
fully shown in 64x64.

Instead we should calculate logicalSize for the overlays,
so we place them in the actual corners of the preview. So
if preview is 64x64, we use that value instead to place the
overlays.