dolphinview: Add widget parameter to elided QToolTip::showText
Without this parameter the QToolTip can not always determine what
is the parent, so it will create a toplevel window in those cases.

Make sure we have dolphinview itself as a widget there, so QToolTip
and thus QtWayland can resolve what is the parent. This
avoids creating random windows when opening a tooltip.

BUG:500009