Fix crash when opening new tab with search
When new tab is opened and it has a search started,
it tries to set the location URL to a navigator that
is not yet necessarily assigned in memory: This is because the new tab is not active.

The locationUrl will be changed properly when tab is activated.

BUG:500101


(cherry picked from commit 16f413ccd786f4ab37a3ea931624a76d5434fe2c)

00e679a4 Fix crash when opening new tab with search
438a0616 Add comment for clarification

Co-authored-by: Akseli Lahtinen <akselmo@akselmo.dev>