DolphinStatusBar: Fix premature text squeezing
The text would get squeezed and elided prematurely due to some clipping
calculations and paddings created for the clipping.

This slightly modifies the clipping code, that we extend the painted
area and then clip off the excess, instead of moving the area to
specific place and then clipping off the excess.

This also makes sure the text width accounts for one extra character
plus the clippings we do, so that text will always fit the statusbar.

BUG: 503164