DolphinView: Don't end anchored selection in updateSelectionState
Do not end anchored selection when selection state is updated, such as
when Dolphin is started with --select "thing.txt" flag.

If we end here, and user wants to press shift to select from
current item to X items, this would cause the anchored selection
to start from the clicked item, not from the first selected item.

BUG: 503221