Check if the item supports sequencing before looking for sequence pixmaps
This patch checks the item first if the plugin it's using for it's previews even supports sequences. If not, we can then safely ignore the hover action for polling more sequences and avoid unnecessary work.

The sequencing support is false by default, so we set it true only when we know it's possible.