dolphinview: Add a dynamic view option
Added a 'dynamic view' option, which allows to switch from a 'compact' or 'details' view to an 'icons' view if most of the files in the directory are images or videos. It reverts to the previous view mode when we switch to a directory which doesn't meet that criteria.

The view mode is only changed once so users don't have to undo that for specific folders when they don't want icon view.

A setting is added in the "Display style" section of the general view setting page.

BUG: 491139