KItemListView: doLayout without animation when editing roles
If user presses delete and then F2 in very quick succession, the
inline rename field will get stuck in place, due to the animation
moving the actual item in different place.

It's better to just refresh the layout without any animations to
make sure the rename field is in correct place, and doesn't stay
stuck in old location.

BUG: 497125