New selection effects
This adds a new selection effect that is similar to what we have in QtQuick file item views.

There are also changes to some usability: Instead of only the icon and text being the clickable area in icon and details mode, the whole selection is now the clickable area.

Otherwise the usability should stay the same, it's mostly a visual change.

See also: https://invent.kde.org/teams/vdg/issues/-/issues/94