DolphinTabWidget: Fix segfault when restoring session
When a session is being restored, DolphinTabPage::setSplitViewEnabled()
is called for each tab. That function assumes the tab page's
m_navigatorsWidget is a valid pointer, which is true only for an active
tab. That causes a segmentation fault when Dolphin is launched from the
command line with more than two URLs and the --split flag.

This commit fixes this bug by connecting the
DolphinNavigatorsWidgetAction to each tab page before calling
restoreState() in DolphinTabWidget::readProperties().

BUG: 462046