KItemListView: Check if animation is already started before starting new one
When user changes the zoom rapidly, we start the animation multiple
times, which causes delays with the animation and the icon size
is wrong for a while until user interaction refreshes the view.

This change runs the animation only when it's been finished, and
avoids the issue with the refresh. Instead in those situations,
we know that user is doing something really rapidly, so we
should just set the icon size to what it needs to be.

BUG: 508352