KStandardItemListWidget: Use styleOption().iconSize
We should use styleOption().iconSize instead of iconSize(), since
the styleOption returns always the correct size depending on
if we are showing previews or not.

In DolphinItemListView we set the iconSize based on the
previewSize or iconSize, and the styleOption().iconSize is
whichever value it is here.

KItemListWidget::iconSize always returns the settings.iconSize

BUG: 508352