Double scroll distance
This commit doubles the distance a single scroll tick will scroll
in the main view.

While fixing a bug and improving scroll behaviour in
8e3addb7e73122a4c89ef347b03f714ff75a253a, the scroll distance was
effectively reduced to a third of the original value. Most users
seem to be okay with this, but there is some disgruntlement. Even
though the new scroll speed is consistent with KWrite, it is way
slower than e.g. Firefox.

In a perfect world we would have a similar scroll distance in every
scroll area so users could predict how much they have to scroll to
see a certain part of the content, however we are not there yet.
This commit aims to find a middle ground between our own definition
of scroll distance and user expectation.

BUG: 508290

(cherry picked from commit 76b2443130955a7a830d06ef23f647950ad48507)