dolphinview: position the focus after the last removed file
Previously, the code could misposition the selection after removal if
the last file in the view was removed and several files were selected.

And since it didn't clear the selection, the previous selection
influence the selection.

Instead clear the selection and position only the focus, either on the
next item after the last index, or the last item still in the view if the last
item was removed.

BUG: 504490