Compile only one autogenerated dolphin settings class
dolphin general settings was built into both the dolphinprivate and
dolphinstatic, a library which links dolphin private. This is
problematic as GeneralSettings::self now points to two different places.

The reason it was being compiled twice was that we need it was private
within dolphinrpivate, but this is fixable.

general_settings was also included in the kcms explicitly which we don't
need.