dolphinview: On shift+middle-click, open file with third app for the open with list
Added the feature to open the selected file with the third application from the associated applications list on shift and middle-click. Generalized the middle-click slot to open the file with the app on a specified index, since the behavior should be almost the same for those two actions.

Similar to: d27ee07de7558470ef7b497fbe3d3504ce7cad07

BUG: 510237