Revert "Avoid implicitly selecting items"
This reverts commit 122fee5625f0285ec4ebda79162c72390989eb2a.

This behavior change was well-intentioned, but has significant usability
and speed drawbacks that have not been addressed:
- Keyboard-driven folder manipulation became slower
- Unexpected behavioral differences when opening files with the pointer
  compared to when opening them with the keyboard introduced
  inconsistency and cognitive load
- Unexpected opening of selection mode during fast operation introduced
  the potential for confusion and additional errors
- Dolphin's behavior became inconsistent with that of other file
  managers users may be accustomed to

The bug tracker, discuss.kde.org, and Reddit are full of complaints
about this change. It's been a year now; I think it's clear that many
Dolphin users have not gotten used to and accepted it. I have to count
myself as one of them, I'm afraid. I've tried to get used to it for a
year, but I just have not been able to.

I don't believe the benefits of this change outweigh the drawbacks, so
let's revert it.

24d859cf19e90fa22ed687b36a68231625c1bd80 was explicitly mentioned as a
thing that should also be reverted in this case, but it's already been
done.

BUG: 494125
BUG: 511966
CCBUG: 424723
CCBUG: 492404
FIXED-IN: 25.12.0