ViewProperties: Avoid tempfile flooding
This patch sets the tempFile to be part of the ViewProperties scope,
so that we do not leave any temporary files littered around whenever
ViewProperties are closed. This fixes the issue with /tmp/ being
flooded with Dolphin temporary files.

It also utilizes the same tempfile until it does not need it anymore,
avoiding creating new ones.

We also check that we can actually open the temporary file before
working on it, since open() is nodiscard.

BUG: 508402